Basic communication of client to server
In this commit, we send the whole message typed in the shell client to the
server.  This is repeated (with a new net.Conn) with each command typed in the
command line.

Right now the receving server just does a dummy action (it prints out what it
got).  Next step is parsing.